Update README.md
# TestAndroid
Platform game (Xamarin/MonoGame)

This is my first experience in game development and also this is a present for my niece. It's a pixel art style platform game for Android. I use Xamarin and MonoGame/XNA for this project. The Game support touch screen, gamepad and keyboard. And so on...